Created a route to tell the backend that a task was completed + a few other fixes

+++ src/controllers/TaskController.js
@@ -1,6 +1,7 @@
 import connection from '../databaseConnection.js'
 import token from '../token.js'
 import jwt from 'jsonwebtoken'
+
 const dateGex = new RegExp('^\\d{4}\\-(0?[1-9]|1[012])\\-(0?[1-9]|[12][0-9]|3[01])$')
 
 export const getTaskById = (req, res) => {
@@ -23,9 +24,14 @@ export const getTaskById = (req, res) => {
 
 export const addTask = async (req, res) => {
     const {ownerToken, parentTaskId, startTime, deadline, description, type} = req.body
-    const decoded= await jwt.verify(ownerToken, token)
-    const ownerId=decoded.userId
-    console.log(ownerId)
+    if (!ownerToken) {
+        res.status(400).json({
+            message: 'You need to send an ownerToken with your request'
+        })
+        return
+    }
+    const decoded = await jwt.verify(ownerToken, token)
+    const ownerId = decoded.userId
     if (type !== 'temp') {
         if (!dateGex.test(deadline)) {
             res.status(400).json({
@@ -90,8 +96,66 @@ export const addTask = async (req, res) => {
 
 }
 
+
+export const updateCompletedStatus = (req, res) => {
+    connection.query(`SELECT userId, parentTaskId FROM tasks WHERE id=${req.params.taskId}`, async (err, rows) => {
+        if (err) {
+            res.status(500).json({
+                message: 'There was a server error when fetching the userId for this task'
+            })
+            throw err
+        }
+        if (!req.body.ownerToken) {
+            res.status(400).json({
+                message: 'You need to send an ownerToken with your request'
+            })
+            return
+        }
+        const decoded = await jwt.verify(req.body.ownerToken, token)
+        const ownerId = decoded.userId
+        if (ownerId !== rows[0].userId) {
+            res.status(400).json({
+                message: 'The id of the user doesn`t match with the id of the owner of this task'
+            })
+            return
+        } else {
+            const parentId = rows[0].parentTaskId
+            connection.query(`UPDATE tasks SET completed=${req.body.completed} WHERE id=${req.params.taskId}`, (err) => {
+                if (err) {
+                    res.status(500).json({
+                        message: 'There was an error when trying to update the complete status of the task'
+                    })
+                    throw err
+                }
+                if (req.body.completed) {
+                    connection.query(`SELECT (SELECT count(*) FROM tasks WHERE parentTaskId=${parentId} AND completed=1)=(SELECT count(*) FROM tasks WHERE parentTaskId=${parentId}) as status`, (err, rows) => {
+                        if (err) {
+                            res.status(500).json({
+                                message: 'There was an error when trying to check the complete status of all the subtasks'
+                            })
+                            throw err
+                        }
+                        res.json({
+                            message: 'Task successfully updated',
+                            parentCompleteness: rows[0].status
+                        })
+                    })
+                } else {
+                    connection.query(`UPDATE tasks SET completed=false WHERE id=${parentId}`, (err) => {
+                        if (err) {
+                            res.status(500).json({
+                                message: 'There was a server error when trying to update the complete status of the parent'
+                            })
+                            throw err
+                        }
+                    })
+                }
+            })
+        }
+    })
+}
+
 export const updateTask = (req, res) => {
-    console.log(req.body)
     let updateFields = []
     for (const field of req.body.columns.split(" ")) {
         if (!['completed', 'userId'].includes(field)) {
@@ -100,33 +164,80 @@ export const updateTask = (req, res) => {
             updateFields.push(`${field} = ${req.body[field]}`)
         }
     }
-    connection.query(`UPDATE tasks SET ${updateFields} WHERE id=${req.params.taskId}`, (err, result) => {
+    connection.query(`SELECT userId FROM tasks WHERE id=${req.params.taskId}`, async (err, rows) => {
         if (err) {
             res.status(500).json({
-                message: 'There was an error when trying to update the task'
+                message: 'There was a server error when fetching the userId for this task'
             })
             throw err
+        }
+        if (!req.body.ownerToken) {
+            res.status(400).json({
+                message: 'You need to send an ownerToken with your request'
+            })
+            return
+        }
+        const decoded = await jwt.verify(req.body.ownerToken, token)
+        const ownerId = decoded.userId
+        if (ownerId !== rows[0].userId) {
+            res.status(400).json({
+                message: 'The id of the user doesn`t match with the id of the owner of this task'
+            })
+            return
         } else {
-            res.json({
-                message: 'The task was updated successfully'
+            connection.query(`UPDATE tasks SET ${updateFields} WHERE id=${req.params.taskId}`, (err, result) => {
+                if (err) {
+                    res.status(500).json({
+                        message: 'There was an error when trying to update the task'
+                    })
+                    throw err
+                } else {
+                    res.json({
+                        message: 'The task was updated successfully'
+                    })
+                }
             })
         }
     })
 }
 
 export const deleteTask = (req, res) => {
-    connection.query(`DELETE FROM tasks WHERE id=${req.params.taskId}`, (err) => {
+    connection.query(`SELECT userId FROM tasks WHERE id=${req.params.taskId}`, async (err, rows) => {
         if (err) {
             res.status(500).json({
-                message: 'There was an error when trying to delete the task'
+                message: 'There was a server error when fetching the userId for this task'
             })
             throw err
+        }
+        if (!req.body.ownerToken) {
+            res.status(400).json({
+                message: 'You need to send an ownerToken with your request'
+            })
+            return
+        }
+        const decoded = await jwt.verify(req.body.ownerToken, token)
+        const ownerId = decoded.userId
+        if (ownerId !== rows[0].userId) {
+            res.status(400).json({
+                message: 'The id of the user doesn`t match with the id of the owner of this task'
+            })
+            return
         } else {
-            res.json({
-                message: 'The task was deleted successfully'
+            connection.query(`DELETE FROM tasks WHERE id=${req.params.taskId}`, (err) => {
+                if (err) {
+                    res.status(500).json({
+                        message: 'There was an error when trying to delete the task'
+                    })
+                    throw err
+                } else {
+                    res.json({
+                        message: 'The task was deleted successfully'
+                    })
+                }
             })
         }
     })
+
 }
 
 export const getSubtasks = (req, res) => {
@@ -150,7 +261,8 @@ export const getSubtasks = (req, res) => {
                     description: row.description,
                     startTime: row.startTime,
                     deadline: row.deadline,
-                    completed: row.completed === 1
+                    completed: row.completed === 1,
+                    failed: row.failed === 1
                 }
             })
             res.json({
